functionaliteit toegevoegd om score aan te passen van Album

+++ src/main/java/com/example/jpamuziek/records/ScoreFormRecord.java
@@ -2,5 +2,5 @@ package com.example.jpamuziek.records;
 
 import org.hibernate.validator.constraints.Range;
 
-public record ScoreFormRecord(@Range(min = 0, max = 10) long score) {
+public record ScoreFormRecord(@Range(min = 0, max = 10) int score) {
 }
